+++ packages/nuget-client/IApiKeyRetriever.cs
@@ -1,6 +1,0 @@
-namespace AT.Brevgenerator.Klient;
-
-public interface IApiKeyRetriever
-{
-    Task<string> RetrieveApiKeyAsync();
-}